简体字转化软件/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "word-traditional-to-simplified"
version = "1.0.0"
description = "Word文档繁体转简体转换工具"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyQt6>=6.4.0",
    "python-docx>=0.8.11",
    "opencc-python-reimplemented>=0.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "hypothesis>=6.0.0",
    "pytest-qt>=4.0.0",
    "pyinstaller>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.hypothesis]
profile = "default"

[project.scripts]
word-converter = "src.main:main"
